Fix build errors and mappings

+++ L4D2VR/vr.cpp
@@ -13,7 +13,7 @@
 #include <thread>
 #include <type_traits>
 #include <algorithm>
-#include <d3d9_vr.h>
+#include "../dxvk/src/d3d9/d3d9_vr.h"
 
 VR::VR(Game *game) 
 {
@@ -93,7 +93,10 @@ VR::VR(Game *game)
     //const vr::HmdVector2_t mouseScaleHUD = {windowWidth, windowHeight};
     //m_Overlay->SetOverlayMouseScale(m_HUDHandle, &mouseScaleHUD);
 
-    const vr::HmdVector2_t mouseScaleMenu = {m_RenderWidth, m_RenderHeight};
+    const vr::HmdVector2_t mouseScaleMenu = {
+        static_cast<float>(m_RenderWidth),
+        static_cast<float>(m_RenderHeight)
+    };
     m_Overlay->SetOverlayCurvature(m_MainMenuHandle, 0.15f);
     m_Overlay->SetOverlayMouseScale(m_MainMenuHandle, &mouseScaleMenu);
